# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "互动教学平台 API",
    "description": "Interactive Teaching Platform Backend API Specification",
    "version": "1.0.0",
    "contact": {
      "name": "Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    },
    {
      "url": "https://api.yourapp.com/api",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "获取课程列表",
        "description": "获取所有课程,支持筛选",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "搜索关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "课程状态",
            "schema": {
              "type": "string",
              "enum": ["Open", "Closed", "Coming Soon"]
            }
          },
          {
            "name": "semester",
            "in": "query",
            "description": "学期筛选",
            "schema": {
              "type": "string",
              "example": "2025 Sem1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取课程列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Course"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "创建课程",
        "description": "创建新课程",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "课程创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Course"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "获取单个课程",
        "description": "根据课程ID获取课程详情",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "课程ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取课程详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Course"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Courses"],
        "summary": "更新课程",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "课程更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Course"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "删除课程",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Course deleted successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用JWT token进行身份验证"
      }
    },
    "schemas": {
...
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "course_1"
          },
          "code": {
            "type": "string",
            "example": "COMP1001"
          },
          "title": {
            "type": "string",
            "example": "Introduction to Programming"
          },
          "status": {
            "type": "string",
            "enum": ["Open", "Closed", "Coming Soon"],
            "example": "Open"
          },
          "schedule": {
            "type": "string",
            "example": "Mon 18:00-21:00"
          },
          "students": {
            "type": "string",
            "example": "45"
          },
          "year": {
            "type": "string",
            "example": "2025"
          },
          "semester": {
            "type": "string",
            "enum": ["Sem1", "Sem2", "Summer Term", "Winter Term"],
            "example": "Sem1"
          },
          "weekday": {
            "type": "string",
            "enum": ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"],
            "example": "Monday"
          },
          "classTime": {
            "type": "string",
            "example": "18:00-21:00"
          },
          "capacity": {
            "type": "string",
            "example": "50"
          }
        }
      },
      "CourseInput": {
        "type": "object",
        "required": ["code", "title", "status"],
        "properties": {
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["Open", "Closed", "Coming Soon"]
          },
          "students": {
            "type": "string"
          },
          "year": {
            "type": "string"
          },
          "semester": {
            "type": "string"
          },
          "weekday": {
            "type": "string"
          },
          "classTime": {
            "type": "string"
          },
          "capacity": {
            "type": "string"
          }
        }
      },
...
    },
...
  }
}